// src/api/generated/models/companyApiKeyScopesItem.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type CompanyApiKeyScopesItem = typeof CompanyApiKeyScopesItem[keyof typeof CompanyApiKeyScopesItem];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CompanyApiKeyScopesItem = {
  manageMembers: 'manageMembers',
  manageMemberPermissions: 'manageMemberPermissions',
  manageCompany: 'manageCompany',
  viewMembers: 'viewMembers',
  viewPartners: 'viewPartners',
  managePartners: 'managePartners',
  viewProducts: 'viewProducts',
  manageProducts: 'manageProducts',
  viewRequests: 'viewRequests',
  manageRequests: 'manageRequests',
  viewQuotes: 'viewQuotes',
  manageQuotes: 'manageQuotes',
  viewInvoices: 'viewInvoices',
  manageInvoices: 'manageInvoices',
  viewPayments: 'viewPayments',
  managePayments: 'managePayments',
  confirmPayments: 'confirmPayments',
  viewShippingInvoices: 'viewShippingInvoices',
  manageShippingInvoices: 'manageShippingInvoices',
  viewConsolidations: 'viewConsolidations',
  manageConsolidations: 'manageConsolidations',
  viewNotifications: 'viewNotifications',
  viewTrace: 'viewTrace',
  manageIntegrations: 'manageIntegrations',
} as const;